Structure preferences to provider
