+++ src/containers/StructurePage/StructureBanner.tsx
@@ -30,6 +30,7 @@ import {
 import { styled } from "@ndla/styled-system/jsx";
 import { NodeType } from "@ndla/types-taxonomy";
 import AddNodeDialogContent from "./AddNodeDialogContent";
+import { usePreferences } from "./PreferencesProvider";
 import { DialogCloseButton } from "../../components/DialogCloseButton";
 import { TAXONOMY_ADMIN_SCOPE } from "../../constants";
 import { useSession } from "../Session/SessionProvider";
@@ -67,43 +68,28 @@ const ButtonsWrapper = styled("div", {
 });
 
 interface Props {
-  setShowFavorites: (checked: boolean) => void;
-  showFavorites: boolean;
-  setShowLmaSubjects: (checked: boolean) => void;
-  setShowDaSubjects: (checked: boolean) => void;
-  setShowSaSubjects: (checked: boolean) => void;
-  showLmaSubjects: boolean;
-  showDaSubjects: boolean;
-  showSaSubjects: boolean;
   nodeType: NodeType;
   hasLmaSubjects: boolean;
   hasDaSubjects: boolean;
   hasSaSubjects: boolean;
-  showQuality: boolean;
-  setShowQuality: (checked: boolean) => void;
-  showMatomoStats: boolean;
-  setShowMatomoStats: (checked: boolean) => void;
 }
 
-const StructureBanner = ({
-  setShowFavorites,
-  showFavorites,
-  setShowLmaSubjects,
-  setShowDaSubjects,
-  setShowSaSubjects,
-  showLmaSubjects,
-  showDaSubjects,
-  showSaSubjects,
-  nodeType,
-  hasLmaSubjects,
-  hasDaSubjects,
-  hasSaSubjects,
-  showQuality,
-  setShowQuality,
-  showMatomoStats,
-  setShowMatomoStats,
-}: Props) => {
+const StructureBanner = ({ nodeType, hasLmaSubjects, hasDaSubjects, hasSaSubjects }: Props) => {
   const [addSubjectDialogOpen, setAddSubjectDialogOpen] = useState(false);
+  const {
+    showFavorites,
+    setShowFavorites,
+    showLmaSubjects,
+    setShowLmaSubjects,
+    showDaSubjects,
+    setShowDaSubjects,
+    showSaSubjects,
+    setShowSaSubjects,
+    showQuality,
+    setShowQuality,
+    showMatomoStats,
+    setShowMatomoStats,
+  } = usePreferences();
 
   const { t } = useTranslation();
   const { userPermissions } = useSession();
